refactor: move error button and loader to features
